+++ src/game/tests.rs
@@ -589,6 +589,7 @@ mod mana_triggers;
 mod mana_vault;
 mod meekstone;
 mod menace;
+mod mishras_war_machine;
 mod modal_effects;
 mod morbid;
 mod multi_block;